Resources: Prefabs in FBX Importer #147 Asset selection mechanism reworked Icon placeholder added for Effect resources Icon render for the prefabs

+++ engine/includes/commandbuffer.h
@@ -28,9 +28,9 @@ public:
 public:
     virtual void                clearRenderTarget           (bool clearColor = true, const Vector4 &color = Vector4(), bool clearDepth = true, float depth = 1.0f);
 
-    virtual void                drawMesh                    (const Matrix4 &model, Mesh *mesh, uint32_t surface = 0, uint8_t layer = ICommandBuffer::DEFAULT, MaterialInstance *material = nullptr);
+    virtual void                drawMesh                    (const Matrix4 &model, Mesh *mesh, uint32_t surface = 0, uint32_t layer = ICommandBuffer::DEFAULT, MaterialInstance *material = nullptr);
 
-    virtual void                drawMeshInstanced           (const Matrix4 *models, uint32_t count, Mesh *mesh, uint32_t surface = 0, uint8_t layer = ICommandBuffer::DEFAULT, MaterialInstance *material = nullptr, bool particle = false);
+    virtual void                drawMeshInstanced           (const Matrix4 *models, uint32_t count, Mesh *mesh, uint32_t surface = 0, uint32_t layer = ICommandBuffer::DEFAULT, MaterialInstance *material = nullptr, bool particle = false);
 
     virtual void                setRenderTarget             (const TargetBuffer &target, const RenderTexture *depth = nullptr);
