+++ engine/includes/components/directlight.h
@@ -16,7 +16,7 @@ public:
 
     ~DirectLight                ();
 
-    void                        draw                    (ICommandBuffer &buffer, int8_t layer);
+    void                        draw                    (ICommandBuffer &buffer, uint32_t layer);
 
     Vector4                    &normalizedDistance      ();
