+++ engine/includes/components/textrender.h
@@ -33,7 +33,7 @@ public:
 public:
     TextRender          ();
 
-    void                draw                (ICommandBuffer &buffer, int8_t layer);
+    void                draw                (ICommandBuffer &buffer, uint32_t layer);
 
     Mesh               *mesh                () const;
